refactor(all): rename base to core
