+++ .env
@@ -1,5 +1,5 @@
 # version
-GOLANG_VERSION=1.19.3
+GOLANG_VERSION=1.21
 K6_VERSION=0.44.0
 
 # service
