+++ internal/external/external.go
@@ -12,7 +12,7 @@ import (
 	"github.com/instill-ai/controller-vdp/config"
 	"github.com/instill-ai/controller-vdp/pkg/logger"
 
-	mgmtPB "github.com/instill-ai/protogen-go/base/mgmt/v1alpha"
+	mgmtPB "github.com/instill-ai/protogen-go/core/mgmt/v1alpha"
 	connectorPB "github.com/instill-ai/protogen-go/vdp/connector/v1alpha"
 	pipelinePB "github.com/instill-ai/protogen-go/vdp/pipeline/v1alpha"
 	etcdv3 "go.etcd.io/etcd/client/v3"
